Module Plplot

PLplot - A library for creating scientific plots

Note: The API for the Plot and Quick_plot modules is not yet fixed and may change between now and the 5.10.0 PLplot release.

The standard PLplot API is not expected to change.

module Plot : sig ... end

module Quick_plot : sig ... end

The standard PLplot API

See the main PLplot documentation for the description and documentation of these functions.

type plplot_axis_type =

| PL_X_AXIS

| PL_Y_AXIS

| PL_Z_AXIS

and plplot3d_style_enum =

| PL_DIFFUSE

| PL_DRAW_LINEX

| PL_DRAW_LINEY

| PL_DRAW_LINEXY

| PL_MAG_COLOR

| PL_BASE_CONT

| PL_TOP_CONT

| PL_SURF_CONT

| PL_DRAW_SIDES

| PL_FACETED

| PL_MESH

and plplot3d_style = plplot3d_style_enum list

and plplot_bin_enum =

| PL_BIN_DEFAULT

| PL_BIN_CENTRED

| PL_BIN_NOEXPAND

| PL_BIN_NOEMPTY

and plplot_bin_style = plplot_bin_enum list

and plplot_hist_enum =

| PL_HIST_DEFAULT

| PL_HIST_NOSCALING

| PL_HIST_IGNORE_OUTLIERS

| PL_HIST_NOEXPAND

| PL_HIST_NOEMPTY

and plplot_hist_style = plplot_hist_enum list

and plplot_run_level_enum =

| PL_UNINITIALIZED

| PL_INITIALIZED

| PL_VIEWPORT_DEFINED

| PL_WORLD_COORDINATES_DEFINED

and plplot_run_level = plplot_run_level_enum

and plplot_position_enum =

| PL_POSITION_LEFT

| PL_POSITION_RIGHT

| PL_POSITION_TOP

| PL_POSITION_BOTTOM

| PL_POSITION_INSIDE

| PL_POSITION_OUTSIDE

| PL_POSITION_VIEWPORT

| PL_POSITION_SUBPAGE

and plplot_position_opt = plplot_position_enum list

and plplot_legend_enum =

| PL_LEGEND_NONE

| PL_LEGEND_COLOR_BOX

| PL_LEGEND_LINE

| PL_LEGEND_SYMBOL

| PL_LEGEND_TEXT_LEFT

| PL_LEGEND_BACKGROUND

| PL_LEGEND_BOUNDING_BOX

| PL_LEGEND_ROW_MAJOR

and plplot_legend_opt = plplot_legend_enum list

and plplot_colorbar_enum =

| PL_COLORBAR_LABEL_LEFT

| PL_COLORBAR_LABEL_RIGHT

| PL_COLORBAR_LABEL_TOP

| PL_COLORBAR_LABEL_BOTTOM

| PL_COLORBAR_IMAGE

| PL_COLORBAR_SHADE

| PL_COLORBAR_GRADIENT

| PL_COLORBAR_CAP_NONE

| PL_COLORBAR_CAP_LOW

| PL_COLORBAR_CAP_HIGH

| PL_COLORBAR_SHADE_LABEL

| PL_COLORBAR_ORIENT_RIGHT

| PL_COLORBAR_ORIENT_TOP

| PL_COLORBAR_ORIENT_LEFT

| PL_COLORBAR_ORIENT_BOTTOM

| PL_COLORBAR_BACKGROUND

| PL_COLORBAR_BOUNDING_BOX

and plplot_colorbar_opt = plplot_colorbar_enum list

and plplot_fci_family_enum =

| PL_FCI_FAMILY_UNCHANGED

| PL_FCI_SANS

| PL_FCI_SERIF

| PL_FCI_MONO

| PL_FCI_SCRIPT

| PL_FCI_SYMBOL

and plplot_fci_style_enum =

| PL_FCI_STYLE_UNCHANGED

| PL_FCI_UPRIGHT

| PL_FCI_ITALIC

| PL_FCI_OBLIQUE

and plplot_fci_weight_enum =

| PL_FCI_WEIGHT_UNCHANGED

| PL_FCI_MEDIUM

| PL_FCI_BOLD

and plplot_draw_mode_enum =

| PL_DRAWMODE_UNKNOWN

| PL_DRAWMODE_DEFAULT

| PL_DRAWMODE_REPLACE

| PL_DRAWMODE_XOR

val pl_setcontlabelformat : int -> int -> unit

val pl_setcontlabelparam : float -> float -> float -> int -> unit

val pladv : int -> unit

val plarc : float -> float -> float -> float -> float -> float -> float -> bool -> unit

val plaxes : float -> float -> string -> float -> int -> string -> float -> int -> unit

val plbin : float array -> float array -> plplot_bin_style -> unit

val plbtime : float -> int * int * int * int * int * float

val plbop : unit -> unit

val plbox : string -> float -> int -> string -> float -> int -> unit

val plbox3 : string -> string -> float -> int -> string -> string -> float -> int -> string -> string -> float -> int -> unit

val plcalc_world : float -> float -> float * float * int

val plclear : unit -> unit

val plcol0 : int -> unit

val plcol1 : float -> unit

val plconfigtime : float -> float -> float -> int -> bool -> int -> int -> int -> int -> int -> float -> unit

val plcpstrm : int -> bool -> unit

val plctime : int -> int -> int -> int -> int -> float -> float

val plend : unit -> unit

val plend1 : unit -> unit

val plenv : float -> float -> float -> float -> int -> int -> unit

val plenv0 : float -> float -> float -> float -> int -> int -> unit

val pleop : unit -> unit

val plerrx : float array -> float array -> float array -> unit

val plerry : float array -> float array -> float array -> unit

val plfamadv : unit -> unit

val plfill : float array -> float array -> unit

val plfill3 : float array -> float array -> float array -> unit

val plflush : unit -> unit

val plfont : int -> unit

val plfontld : int -> unit

val plgchr : unit -> float * float

val plgcmap1_range : unit -> float * float

val plgcol0 : int -> int * int * int

val plgcol0a : int -> int * int * int * float

val plgcolbg : unit -> int * int * int

val plgcolbga : unit -> int * int * int * float

val plgcompression : unit -> int

val plgdev : unit -> string

val plgdidev : unit -> float * float * float * float

val plgdiori : unit -> float

val plgdiplt : unit -> float * float * float * float

val plgdrawmode : unit -> plplot_draw_mode_enum

val plgfci : unit -> int64

val plgfam : unit -> int * int * int

val plgfnam : unit -> string

val plgfont : unit -> int * int * int

val plglevel : unit -> plplot_run_level

val plgpage : unit -> float * float * int * int * int * int

val plgra : unit -> unit

val plgradient : float array -> float array -> float -> unit

val plgspa : unit -> float * float * float * float

val plgstrm : unit -> int

val plgver : unit -> string

val plgvpd : unit -> float * float * float * float

val plgvpw : unit -> float * float * float * float

val plgxax : unit -> int * int

val plgyax : unit -> int * int

val plgzax : unit -> int * int

val plhist : float array -> float -> float -> int -> plplot_hist_style -> unit

val plhlsrgb : float -> float -> float -> float * float * float

val plinit : unit -> unit

val pljoin : float -> float -> float -> float -> unit

val pllab : string -> string -> string -> unit

val plcolorbar : plplot_colorbar_opt -> plplot_position_opt -> float -> float -> float -> float -> int -> int -> int -> float -> float -> int -> float -> plplot_colorbar_opt array -> string array -> string array -> float array -> int array -> float array array -> float * float

val pllegend : plplot_legend_opt -> plplot_position_opt -> float -> float -> float -> int -> int -> int -> int -> int -> plplot_legend_opt array -> float -> float -> float -> float -> int array -> string array -> int array -> int array -> float array -> float array -> int array -> int array -> float array -> int array -> float array -> int array -> string array -> float * float

val pllightsource : float -> float -> float -> unit

val plline : float array -> float array -> unit

val plline3 : float array -> float array -> float array -> unit

val pllsty : int -> unit

val plmesh : float array -> float array -> float array array -> plplot3d_style -> unit

val plmeshc : float array -> float array -> float array array -> plplot3d_style -> float array -> unit

val plmkstrm : unit -> int

val plmtex : string -> float -> float -> float -> string -> unit

val plmtex3 : string -> float -> float -> float -> string -> unit

val plot3d : float array -> float array -> float array array -> plplot3d_style -> bool -> unit

val plot3dc : float array -> float array -> float array array -> plplot3d_style -> float array -> unit

val plpat : int array -> int array -> unit

val plpath : int -> float -> float -> float -> float -> unit

val plpoin : float array -> float array -> int -> unit

val plpoin3 : float array -> float array -> float array -> int -> unit

val plprec : int -> int -> unit

val plpsty : int -> unit

val plptex : float -> float -> float -> float -> float -> string -> unit

val plptex3 : float -> float -> float -> float -> float -> float -> float -> float -> float -> float -> string -> unit

val plrandd : unit -> float

val plreplot : unit -> unit

val plrgbhls : float -> float -> float -> float * float * float

val plschr : float -> float -> unit

val plscmap0 : int array -> int array -> int array -> unit

val plscmap0a : int array -> int array -> int array -> float array -> unit

val plscmap0n : int -> unit

val plscmap1 : int array -> int array -> int array -> unit

val plscmap1a : int array -> int array -> int array -> float array -> unit

val plscmap1l : bool -> float array -> float array -> float array -> float array -> bool array option -> unit

val plscmap1la : bool -> float array -> float array -> float array -> float array -> float array -> bool array option -> unit

val plscmap1n : int -> unit

val plscmap1_range : float -> float -> unit

val plscol0 : int -> int -> int -> int -> unit

val plscol0a : int -> int -> int -> int -> float -> unit

val plscolbg : int -> int -> int -> unit

val plscolbga : int -> int -> int -> float -> unit

val plscolor : int -> unit

val plscompression : int -> unit

val plsdev : string -> unit

val plsdidev : float -> float -> float -> float -> unit

val plsdimap : int -> int -> int -> int -> float -> float -> unit

val plsdiori : float -> unit

val plsdiplt : float -> float -> float -> float -> unit

val plsdiplz : float -> float -> float -> float -> unit

val plseed : int64 -> unit

val plsesc : char -> unit

val plsfam : int -> int -> int -> unit

val plsfci : int64 -> unit

val plsfnam : string -> unit

val plsfont : plplot_fci_family_enum -> plplot_fci_style_enum -> plplot_fci_weight_enum -> unit

val plsmaj : float -> float -> unit

val plsmin : float -> float -> unit

val plsdrawmode : plplot_draw_mode_enum -> unit

val plsori : int -> unit

val plspage : float -> float -> int -> int -> int -> int -> unit

val plspal0 : string -> unit

val plspal1 : string -> bool -> unit

val plspause : bool -> unit

val plsstrm : int -> unit

val plssub : int -> int -> unit

val plssym : float -> float -> unit

val plstar : int -> int -> unit

val plstart : string -> int -> int -> unit

val plstring : float array -> float array -> string -> unit

val plstring3 : float array -> float array -> float array -> string -> unit

val plstripa : int -> int -> float -> float -> unit

val plstripd : int -> unit

val plimage : float array array -> float -> float -> float -> float -> float -> float -> float -> float -> float -> float -> unit

val plstyl : int array -> int array -> unit

val plsurf3d : float array -> float array -> float array array -> plplot3d_style -> float array -> unit

val plsvect : float array -> float array -> bool -> unit

val plsvpa : float -> float -> float -> float -> unit

val plsxax : int -> int -> unit

val plsxwin : int -> unit

val plsyax : int -> int -> unit

val plsym : float array -> float array -> int -> unit

val plszax : int -> int -> unit

val pltext : unit -> unit

val pltimefmt : string -> unit

val plvasp : float -> unit

val plvpas : float -> float -> float -> float -> float -> unit

val plvpor : float -> float -> float -> float -> unit

val plvsta : unit -> unit

val plw3d : float -> float -> float -> float -> float -> float -> float -> float -> float -> float -> float -> unit

val plwidth : float -> unit

val plwind : float -> float -> float -> float -> unit

val plxormod : bool -> bool

val plsetopt : string -> string -> unit

val plMinMax2dGrid : float array array -> float * float

val plcont : float array array -> int -> int -> int -> int -> float array -> unit

val plshade : float array array -> float -> float -> float -> float -> float -> float -> int -> float -> float -> int -> float -> int -> float -> bool -> unit

val plshades : float array array -> float -> float -> float -> float -> float array -> float -> int -> float -> bool -> unit

val plimagefr : float array array -> float -> float -> float -> float -> float -> float -> float -> float -> unit

val plvect : float array array -> float array array -> float -> unit

val plmap : string -> float -> float -> float -> float -> unit

val plmeridians : float -> float -> float -> float -> float -> float -> unit

val plpoly3 : float array -> float array -> float array -> bool array -> bool -> unit

val pltr0 : float -> float -> float * float

val plsvect_reset : unit -> unit

val plg_current_col0 : unit -> int

val plg_current_col1 : unit -> float

val plgwidth : unit -> float

val plgchrht : unit -> float

val plstripc : string -> string -> float -> float -> float -> float -> float -> float -> float -> bool -> bool -> int -> int -> int array -> int array -> string array -> string -> string -> string -> int

val pltr1 : float array -> float array -> float -> float -> float * float

val pltr2 : float array array -> float array array -> float -> float -> float * float

val plset_pltr : (float -> float -> float * float) -> unit

val plunset_pltr : unit -> unit

val plset_mapform : (float -> float -> float * float) -> unit

val plunset_mapform : unit -> unit

val plset_defined : (float -> float -> int) -> unit

val plunset_defined : unit -> unit

val plstransform : (float -> float -> float * float) -> unit

val plunset_transform : unit -> unit

type plplot_grid_method_type =

| PL_GRID_CSA

| PL_GRID_DTLI

| PL_GRID_NNI

| PL_GRID_NNIDW

| PL_GRID_NNLI

| PL_GRID_NNAIDW

type plplot_parse_method_type =

| PL_PARSE_PARTIAL

| PL_PARSE_FULL

| PL_PARSE_QUIET

| PL_PARSE_NODELETE

| PL_PARSE_SHOWALL

| PL_PARSE_OVERRIDE

| PL_PARSE_NOPROGRAM

| PL_PARSE_NODASH

| PL_PARSE_SKIP

val plslabelfunc : (plplot_axis_type -> float -> string) -> unit

val plunset_labelfunc : unit -> unit

val plsabort : (string -> unit) -> unit

val plunset_abort : unit -> unit

val plsexit : (string -> int) -> unit

val plunset_exit : unit -> unit

val plgriddata : float array -> float array -> float array -> float array -> float array -> plplot_grid_method_type -> float -> float array array

val plparseopts : string array -> plplot_parse_method_type list -> unit

Non-standard convenience functions (not part of the standard PLplot API.

val plcalc_device : float -> float -> float * float

plcalc_device x y will give the device position, in normalized device coordinates, of the world coordinates (x, y).

val plfullcanvas : unit -> unit

plfullcanvas () maximizes the plot viewport and window. Dimensions are set to (0.0, 0.0) to (1.0, 1.0).

val plpolyline : float array -> float array -> unit

Draw an unfilled polygon. The only functional difference between this and plline is that this function will close the given polygon, so there is no need to duplicate points to have a closed figure.