Module Plplot.Quick_plot

A module for quick, "throw-away" plots

THIS API IS NOT FIXED AND MAY CHANGE BEFORE THE 5.10.0 RELEASE

val points : ?filename:string -> ?size:(int * int) -> ?device:Plot.plot_device_t -> ?labels:(string * string * string) -> ?names:string list -> ?log:(bool * bool) -> (float array * float array) list -> unit

points xs ys plots the points described by the coordinates xs and ys.

val lines : ?filename:string -> ?size:(int * int) -> ?device:Plot.plot_device_t -> ?labels:(string * string * string) -> ?names:string list -> ?log:(bool * bool) -> (float array * float array) list -> unit

lines xs ys plots the line segments described by the coordinates xs and ys.

val image : ?filename:string -> ?size:(int * int) -> ?device:Plot.plot_device_t -> ?labels:(string * string * string) -> ?log:bool -> ?palette:Plot.color_palette_t -> float array array -> unit

image ?log m plots the image m with a matching colorbar. If log is true then the data in m are assumed to be log10(x) values.

val func : ?filename:string -> ?size:(int * int) -> ?device:Plot.plot_device_t -> ?labels:(string * string * string) -> ?names:string list -> ?symbol:string -> ?step:float -> (float -> float) list -> (float * float) -> unit

func ?point ?step fs (min, max) plots the functions fs from x = min to x = max. step can be used to tighten or coarsen the sampling of plot points.

val shades : ?filename:string -> ?size:(int * int) -> ?device:Plot.plot_device_t -> ?labels:(string * string * string) -> ?log:bool -> ?palette:Plot.color_palette_t -> ?contours:float array -> float array array -> unit

shades ?log ?contours m plots a filled contour/shaded m with a matching colorbar. If log is true then the data in m are assumed to be log10(x) values.